// prisma/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AccountType = {
  DEBIT: 'DEBIT',
  CREDIT: 'CREDIT',
  CASH: 'CASH',
  BANK: 'BANK',
  INVESTMENT: 'INVESTMENT',
  DEPOSIT: 'DEPOSIT',
  DIGITAL: 'DIGITAL',
  SAVINGS: 'SAVINGS'
} as const

export type AccountType = (typeof AccountType)[keyof typeof AccountType]


export const CurrencyType = {
  UAH: 'UAH',
  USD: 'USD',
  EUR: 'EUR'
} as const

export type CurrencyType = (typeof CurrencyType)[keyof typeof CurrencyType]


export const TransactionType = {
  EXPENSE: 'EXPENSE',
  INCOME: 'INCOME',
  TRANSFER: 'TRANSFER'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]